Move some common defs back into core
