+++ ui/core/src/model/layout.ts
@@ -11,6 +11,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import { PanelRef } from './panels';
+
 export type LayoutDefinition = GridDefinition;
 
 export interface GridDefinition {
@@ -29,4 +31,5 @@ export interface GridItemDefinition {
   y: number;
   width: number;
   height: number;
+  content: PanelRef;
 }
