+++ ui/core/src/utils/index.ts
@@ -11,8 +11,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export * from './model';
-export * from './runtime/datasources';
-export * from './runtime/template-variables';
-export * from './runtime/time-range';
-export * from './utils';
+export * from './fetch';
+export * from './memo';
+export * from './panel-refs';
